requirements.txt
streamlit
pytesseract
python-dotenv
PyPDF2
langchain
langchain_google_genai
faiss-cpu
pypdfium2
Pillow
langchain_groq
